+++ src/openfacefx/__init__.py
@@ -60,6 +60,8 @@ from .export_live2d import write_live2d_motion, lipsync_param_ids
 from .export_godot import write_godot_anim
 from .export_gltf import write_gltf, build_gltf
 from .export_vmd import write_vmd, vmd_bytes, DEFAULT_MORPH_MAP
+from .export_livelink import (write_livelink_csv, livelink_csv_string,
+                              LIVELINK_COLUMNS)
 from .export_lip import (write_lip, lip_bytes, skyrim_mapping,
                         SKYRIM_SLOT_MAP)
 from .export_cues import (
@@ -139,6 +141,7 @@ __all__ = [
     "write_live2d_motion", "lipsync_param_ids", "write_godot_anim",
     "write_gltf", "build_gltf",
     "write_vmd", "vmd_bytes", "DEFAULT_MORPH_MAP",
+    "write_livelink_csv", "livelink_csv_string", "LIVELINK_COLUMNS",
     "write_lip", "lip_bytes", "skyrim_mapping", "SKYRIM_SLOT_MAP",
     "dominant_cues", "write_rhubarb_tsv", "write_rhubarb_xml",
     "write_rhubarb_json", "write_moho_dat", "write_pgo",